Work stopping point commit
Made a lot of good progress with infrastructure for allowing the code to run in batch mode while making sure to update the Batch Master sheet. Ready to start work on the main function for analyzing images and happy with how I structured everything up until this point in the pipeline as functions.