Premajor change commit
Did a lot of work this morning just developing the basic features but now I'm seeing I need to change the location of the "matching" feature to call it earlier in the macro to avoid unnecessary repetition. Committing code with the functional feature now in case I needed to roll back after the move.