Work stop point
Got to the point of drawing a box around the region of interest and finding some  XYZs that look like they should have been included are not. Need to go through and figure out if it's the calculation, the box being drawn wrong, or the testing being too strict. First check that it's not because they're out of the included z-range.