Debugging
Fixed two minor issues:
(1) XYZ dataset-image matching was happening outside of the for-loop and using the for-loop i-index. Moved the matching to within the for loop to ensure dataset-image matches occur for every XYZ dataset available.
(2) added code for closeing all .csv windows that are open to avoid clutter and accidental access to the wrong dataframe

Found an issue that will need to be addresses ASAP:
Currently the code is written assuming that the index of each XYZ in the Table as opened is the ID, because this is how it was setup in Python. However, this is not the case in the ImageJ macro. When the .csv is opened the index is just whatever ImageJ assigns and "ID" is treated as a column. This means that if ID 1194 is being called, but the Table is only 500 rows, the call will be out of index.