Major updates completed
At this point code has all of the bare minimum functionality that's needed for beta testing with a full dataset. Will keep the structure the same and only fix minor bugs. Code still seems to have the rounding error but I'm still deciding on the best way to fix this issue.